pyproject.toml
[tool.poetry]
authors = ["Alexander Kuhn-Regnier"]
description = "Analyse budgeting info"
name = "budget"
version = "0.0.1"

[tool.poetry.dependencies]
ipdb = "*"
ipython = "*"
loguru = "*"
matplotlib = "*"
numpy = "*"
pandas = "*"
pandera = "*"
pydantic = "*"
python = "^3.11.4"
pyyaml = "*"
ruff = "*"
scipy = "*"

[tool.ruff]
target-version = "py311"
